chore: refactor tests add vcr mocks
