+++ projects/pgai/tests/vectorizer/extensions/test_sqlalchemy_composite_primary.py
@@ -1,12 +1,13 @@
+from typing import Any
+
 import numpy as np
-from click.testing import CliRunner
 from sqlalchemy import Column, Engine, Text
 from sqlalchemy.orm import DeclarativeBase, Session
 from sqlalchemy.sql import text
 from testcontainers.postgres import PostgresContainer  # type: ignore
 
-from pgai.cli import vectorizer_worker
 from pgai.sqlalchemy import embedding_relationship
+from tests.vectorizer.extensions.utils import run_vectorizer_worker
 
 
 class Base(DeclarativeBase):
@@ -23,24 +24,10 @@ class Author(Base):
     )
 
 
-def run_vectorizer_worker(db_url: str, vectorizer_id: int) -> None:
-    CliRunner().invoke(
-        vectorizer_worker,
-        [
-            "--db-url",
-            db_url,
-            "--once",
-            "--vectorizer-id",
-            str(vectorizer_id),
-            "--concurrency",
-            "1",
-        ],
-        catch_exceptions=False,
-    )
-
-
 def test_vectorizer_composite_key(
-    postgres_container: PostgresContainer, initialized_engine: Engine
+    postgres_container: PostgresContainer,
+    initialized_engine: Engine,
+    vcr_: Any,
 ):
     """Test vectorizer with a composite primary key."""
     db_url = postgres_container.get_connection_url()
@@ -74,11 +61,11 @@ def test_vectorizer_composite_key(
         session.commit()
 
     # Run vectorizer worker
-    run_vectorizer_worker(db_url, 1)
+    with vcr_.use_cassette("test_vectorizer_composite_key.yaml"):
+        run_vectorizer_worker(db_url, 1)
 
     # Verify embeddings were created
     with Session(initialized_engine) as session:
-        # Verify embedding class was created correctly
         assert Author.bio_embeddings.__name__ == "BioEmbeddingsEmbedding"
 
         # Check embeddings exist and have correct properties
